LayoutTests/imported/w3c/web-platform-tests/css/css-values/calc-infinity-nan-computed-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: testComputedValueGreaterOrLowerThan

Harness Error (FAIL), message = ReferenceError: Can't find variable: testComputedValueGreaterOrLowerThan

PASS Property width value 'calc(NaN * 1px)'
PASS Property width value 'calc(NaN * 1%)'
